Fixes #17: Move away from uuids as primary identifiers
I don't know why we chose to use uuids as primary identifiers all those
years ago, but it's impractical when looking at a database when you're
debugging something, I can imagine it's slower (to join, index, filter,
... I have no actual data about this, it's an assumption), it's not the
industry standard, and it's causing issues while upgrading to Symfony 5
because doctrine/dbal no longer supports it, and the implementation
Symfony created for it vastly differs from the previous implementation
by doctrine, while the benefits are... non-existent as far as I'm aware.

This is a breaking change, users of this bundle will need to provide
their own migrations for this change.