Add caching for template discovery
Page loads for application environment edit forms took _several_
seconds. After investigation it was due to the template discovery that
we did to give the end user the option to insert the templates using a
dialog.

We added the cache to the template discovery and used cache tags to tag
them with the `template_help` cache tag. This way we can invalidate them
when a new token is added or one is removed.